+++ wdfn-server/config.py
@@ -42,6 +42,7 @@ DATA_DIR = os.path.join(os.path.dirname(__file__), 'data')
 NWIS_CODE_LOOKUP_FILENAME = 'nwis_lookup.json'
 COUNTRY_STATE_COUNTY_LOOKUP_FILENAME = 'nwis_country_state_lookup.json'
 HUC_LOOKUP_FILENAME = 'huc_lookup.json'
+COOPERATOR_DATA_FILENAME = 'cooperator_data.json'
 
 GA_TRACKING_CODE = ''
 ENABLE_USGS_GA = False