+++ wdfn-server/waterdata/views.py
@@ -7,7 +7,7 @@ from flask import abort, render_template, request, Markup
 
 from . import app, __version__
 from .location_utils import build_linked_data, get_disambiguated_values, rollup_dataseries
-from .utils import construct_url, defined_when, execute_get_request, parse_rdb, execute_cooperator_lookup_request
+from .utils import construct_url, defined_when, execute_get_request, parse_rdb, get_cooperator_data
 
 # Station Fields Mapping to Descriptions
 from .constants import STATION_FIELDS_D
@@ -112,17 +112,9 @@ def monitoring_location(site_no):
             except KeyError:
                 site_owner_state = None
 
-            # get the cooperator data from service
-            # feature toggle; remove 'if/else' when new lookup service is implemented
-            # for now, limit to district codes 20 and 51
-            if app.config['COOPERATOR_LOOKUP_ENABLED'] and (
-                    app.config['COOPERATOR_LOOKUP_ENABLED'] is True or
-                    location_with_values.get('district_cd', {}).get('code') in app.config['COOPERATOR_LOOKUP_ENABLED']):
-                params = 'SiteNumber=' + site_no + app.config['URL_PARAMS_COOPERATOR_LOOKUP']
-                cooperator_lookup_data = execute_cooperator_lookup_request(app.config['SERVICE_ROOT_COOPERATOR_LOOKUP'],
-                                                                           app.config['URL_PATH_COOPERATOR_LOOKUP'], params)
-            else:
-                cooperator_lookup_data = None
+            # grab the cooperator information from json file so that the logos are added to page, if available
+            cooperator_lookup_data = get_cooperator_data(location_with_values.get('district_cd', {}).get('code'),
+                                                         site_no)
 
             if site_owner_state is not None:
                 questions_link_params = {